+++ src/metrics/metrics-events-listener.mjs
@@ -1,4 +1,5 @@
 import { Events as MetricsEvents } from './events.mjs';
+import { EventsConfig } from './events-config.mjs';
 import { Events } from '../event/events.mjs';
 import { State } from '../banner/state.mjs';
 
@@ -7,6 +8,11 @@ export class MetricsEventsListener {
     #eventBus;
     #channelCode;
     #attached;
+    #beforeAttachedQueue = {
+        started: false,
+        events: [],
+        cleanup: [],
+    };
 
     /**
      * @param {MetricsSender} metricsSender
@@ -20,7 +26,75 @@ export class MetricsEventsListener {
         this.#attached = false;
     }
 
-    attach() {
+    collectBeforeAttach() {
+        if (this.#beforeAttachedQueue.started) {
+            return;
+        }
+
+        this.#beforeAttachedQueue.started = true;
+
+        const eventBus = this.#eventBus;
+        const config = new EventsConfig({});
+
+        this.#beforeAttachedQueue.cleanup.push(
+            eventBus.subscribe(Events.ON_BANNER_STATE_CHANGED, ({ banner }) => {
+                if (banner.isEmbed() || State.RENDERED !== banner.state || 1 !== banner.stateCounter) {
+                    return;
+                }
+
+                for (let fingerprint of banner.fingerprints) {
+                    this.#beforeAttachedQueue.events.push({
+                        name: MetricsEvents.BANNER_LOADED,
+                        params: this.#createBaseMetricsParams({ config, fingerprint, banner }),
+                    });
+                }
+            }),
+        );
+
+        this.#beforeAttachedQueue.cleanup.push(
+            eventBus.subscribe(Events.ON_BANNER_FIRST_TIME_SEEN, ({ fingerprint, banner }) => {
+                this.#beforeAttachedQueue.events.push({
+                    name: MetricsEvents.BANNER_DISPLAYED,
+                    params: this.#createBaseMetricsParams({ config, fingerprint, banner }),
+                });
+            }),
+        );
+
+        this.#beforeAttachedQueue.cleanup.push(
+            eventBus.subscribe(Events.ON_BANNER_FIRST_TIME_FULLY_SEEN, ({ fingerprint, banner }) => {
+                this.#beforeAttachedQueue.events.push({
+                    name: MetricsEvents.BANNER_FULLY_DISPLAYED,
+                    params: this.#createBaseMetricsParams({ config, fingerprint, banner }),
+                });
+            }),
+        );
+
+        this.#beforeAttachedQueue.cleanup.push(
+            eventBus.subscribe(Events.ON_BANNER_LINK_CLICKED, ({ fingerprint, banner, target }) => {
+                const params = this.#createBaseMetricsParams({ config, fingerprint, banner });
+                params[config.params.link] = target.href || '';
+
+                this.#beforeAttachedQueue.events.push({
+                    name: MetricsEvents.BANNER_CLICKED,
+                    params: params,
+                });
+            }),
+        );
+
+        this.#beforeAttachedQueue.cleanup.push(
+            eventBus.subscribe(Events.ON_BANNER_AFTER_CLOSE, ({ fingerprint, banner }) => {
+                this.#beforeAttachedQueue.events.push({
+                    name: MetricsEvents.BANNER_CLOSED,
+                    params: this.#createBaseMetricsParams({ config, fingerprint, banner }),
+                });
+            }),
+        );
+    }
+
+    /**
+     * @param {EventsConfig} config
+     */
+    attach(config) {
         if (this.#attached) {
             return;
         }
@@ -28,68 +102,112 @@ export class MetricsEventsListener {
         this.#attached = true;
 
         const eventBus = this.#eventBus;
-        const channelCode = this.#channelCode;
         const metricsSender = this.#metricsSender;
 
         if (!metricsSender.hasAnyReceiver()) {
+            this.#clearBeforeAttachedQueue();
+
             return;
         }
 
-        const createBaseMetricsArgs = (fingerprint, banner) => {
-            let breakpoint = banner.getCurrentBreakpoint(fingerprint.bannerId);
-            breakpoint = null === breakpoint ? 'default' : `${'min' === banner.positionData.breakpointType ? '>=' : '<='}${breakpoint}`;
+        const bannerLoadedEventName = config.events[MetricsEvents.BANNER_LOADED];
+        const bannerDisplayedEventName = config.events[MetricsEvents.BANNER_DISPLAYED];
+        const bannerFullyDisplayedEventName = config.events[MetricsEvents.BANNER_FULLY_DISPLAYED];
+        const bannerClickedEventName = config.events[MetricsEvents.BANNER_CLICKED];
+        const bannerClosedEventName = config.events[MetricsEvents.BANNER_CLOSED];
 
-            return {
-                channel_code: channelCode,
-                banner_id: fingerprint.bannerId,
-                banner_name: fingerprint.bannerName,
-                position_id: fingerprint.positionId,
-                position_code: fingerprint.positionCode,
-                position_name: fingerprint.positionName,
-                campaign_id: fingerprint.campaignId,
-                campaign_code: fingerprint.campaignCode,
-                campaign_name: fingerprint.campaignName,
-                breakpoint: breakpoint,
+        if (this.#beforeAttachedQueue.started) {
+            for (let i in this.#beforeAttachedQueue.events) {
+                const { name, params } = this.#beforeAttachedQueue.events[i];
+
+                if (false === config.events[name]) {
+                    continue;
+                }
+
+                const mappedParams = {};
+
+                for (let paramKey in params) {
+                    mappedParams[config.params[paramKey]] = params[paramKey];
+                }
+
+                metricsSender.send(config.events[name], mappedParams);
             }
-        };
 
-        if (metricsSender.isEventEnabled(MetricsEvents.BANNER_LOADED)) {
+            this.#clearBeforeAttachedQueue();
+        }
+
+        if (false !== bannerLoadedEventName) {
             eventBus.subscribe(Events.ON_BANNER_STATE_CHANGED, ({ banner }) => {
                 if (banner.isEmbed() || State.RENDERED !== banner.state || 1 !== banner.stateCounter) {
                     return;
                 }
 
                 for (let fingerprint of banner.fingerprints) {
-                    metricsSender.send(MetricsEvents.BANNER_LOADED, createBaseMetricsArgs(fingerprint, banner));
+                    metricsSender.send(bannerLoadedEventName, this.#createBaseMetricsParams({ config, fingerprint, banner }));
                 }
             });
         }
 
-        if (metricsSender.isEventEnabled(MetricsEvents.BANNER_DISPLAYED)) {
+        if (false !== bannerDisplayedEventName) {
             eventBus.subscribe(Events.ON_BANNER_FIRST_TIME_SEEN, ({ fingerprint, banner }) => {
-                metricsSender.send(MetricsEvents.BANNER_DISPLAYED, createBaseMetricsArgs(fingerprint, banner));
+                metricsSender.send(bannerDisplayedEventName, this.#createBaseMetricsParams({ config, fingerprint, banner }));
             });
         }
 
-        if (metricsSender.isEventEnabled(MetricsEvents.BANNER_FULLY_DISPLAYED)) {
+        if (false !== bannerFullyDisplayedEventName) {
             eventBus.subscribe(Events.ON_BANNER_FIRST_TIME_FULLY_SEEN, ({ fingerprint, banner }) => {
-                metricsSender.send(MetricsEvents.BANNER_FULLY_DISPLAYED, createBaseMetricsArgs(fingerprint, banner));
+                metricsSender.send(bannerFullyDisplayedEventName, this.#createBaseMetricsParams({ config, fingerprint, banner }));
             });
         }
 
-        if (metricsSender.isEventEnabled(MetricsEvents.BANNER_CLICKED)) {
+        if (false !== bannerClickedEventName) {
             eventBus.subscribe(Events.ON_BANNER_LINK_CLICKED, ({ fingerprint, banner, target }) => {
-                metricsSender.send(MetricsEvents.BANNER_CLICKED, {
-                    ...createBaseMetricsArgs(fingerprint, banner),
-                    link: target.href || '',
-                })
+                const params = this.#createBaseMetricsParams({ config, fingerprint, banner });
+                params[config.params.link] = target.href || '';
+
+                metricsSender.send(bannerClickedEventName, params);
             });
         }
 
-        if (metricsSender.isEventEnabled(MetricsEvents.BANNER_CLOSED)) {
+        if (false !== bannerClosedEventName) {
             eventBus.subscribe(Events.ON_BANNER_AFTER_CLOSE, ({ fingerprint, banner }) => {
-                metricsSender.send(MetricsEvents.BANNER_CLOSED, createBaseMetricsArgs(fingerprint, banner));
+                metricsSender.send(bannerClosedEventName, this.#createBaseMetricsParams({ config, fingerprint, banner }));
             });
         }
+    }
+
+    /**
+     * @param {EventsConfig} config
+     * @param {Fingerprint} fingerprint
+     * @param {Banner} banner
+     *
+     * @return {Object}
+     */
+    #createBaseMetricsParams = ({ config, fingerprint, banner }) => {
+        let breakpoint = banner.getCurrentBreakpoint(fingerprint.bannerId);
+        breakpoint = null === breakpoint ? 'default' : `${'min' === banner.positionData.breakpointType ? '>=' : '<='}${breakpoint}`;
+
+        const params = {};
+        params[config.params.channel_code] = this.#channelCode;
+        params[config.params.banner_id] = fingerprint.bannerId;
+        params[config.params.banner_name] = fingerprint.bannerName;
+        params[config.params.position_id] = fingerprint.positionId;
+        params[config.params.position_code] = fingerprint.positionCode;
+        params[config.params.position_name] = fingerprint.positionName;
+        params[config.params.campaign_id] = fingerprint.campaignId;
+        params[config.params.campaign_code] = fingerprint.campaignCode;
+        params[config.params.campaign_name] = fingerprint.campaignName;
+        params[config.params.breakpoint] = breakpoint;
+
+        return params;
+    };
+
+    #clearBeforeAttachedQueue() {
+        for (let i in this.#beforeAttachedQueue.cleanup) {
+            this.#beforeAttachedQueue.cleanup[i]();
+        }
+
+        this.#beforeAttachedQueue.events = [];
+        this.#beforeAttachedQueue.cleanup = [];
     }
 }